Added electron wrapper to utilize desktop clipboard functionality
PWA-only approach isn't exactly great for clipboards. So, I created an electron
wrapper to facilitate background clipboard watching and such. It will also
be useful for handling pictures etc. This will likely create some sort of issue
down the road, but for now, it's a good start. If clipboard api becomes more
mature on web, then this part can be dropped for a pure PWA.